# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "inkentry-server",
    "description": "Shared memory server for spelunk. Stores decisions, requirements, and context for a team and serves them over HTTP.",
    "contact": {
      "name": "spelunk",
      "url": "https://github.com/spelunk-cloud/spelunk"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}/memory": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "List memory entries for a project, optionally filtered by kind.",
        "operationId": "list_notes",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "Filter by kind (`decision`, `requirement`, `note`, `question`, `handoff`, `intent`).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return (default: 20).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "archived",
            "in": "query",
            "description": "Include archived entries (default: false).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Add a memory entry to a project. The project is auto-created on first write.",
        "description": "The `embedding` field is optional. If omitted and the server's embedder is ready,\nthe server embeds the entry before storage. If neither is available, the\nentry is stored without a vector (text search only, no KNN).\n\nReturns **201** on success. Returns **409** when the new entry is semantically\nclose to one or more existing active entries (similarity ≥ conflict_threshold).\nThe entry is still stored in both cases; the 409 is informational.\nReturns **422** when the entry contains prompt-injection patterns.",
        "operationId": "add_note",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug (e.g. `usercise/spelunk`)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Note created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Embedding dimension mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Note stored but conflicts with existing entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Entry rejected: prompt injection detected"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/memory/batch": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Batch-create memory entries. Idempotent on `external_id`: a live note\nalready carrying the given `external_id` is skipped, not duplicated.",
        "description": "Wire-compatible with the CLI's `BatchPushItem`/`CloudSyncClient::push_batch`\n(the same client cloud-api's `/memory/batch` serves); `spelunk memory push`\nand `spelunk sync` target this route on an OSS team server exactly as they\ndo against cloud-api. Always returns **207** with a per-entry result list;\na request-level validation failure (oversized batch, a title/body over the\nconfigured caps, or an injection match) rejects the whole batch (4xx/422)\nwith nothing stored, before any entry is written.",
        "operationId": "push_memory_batch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug (e.g. `usercise/spelunk`)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "207": {
            "description": "Per-entry outcomes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPushResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (oversized batch, bad field length)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Entry rejected: prompt injection detected"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/projects/{project_id}/memory/harvested-shas": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Return all git commit SHAs stored in note tags for a project.",
        "description": "Each harvested commit is tagged `git:<sha>`. Clients call this endpoint\nto skip commits they have already stored, enabling incremental harvest.",
        "operationId": "harvested_shas",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of harvested git SHAs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarvestedShasResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddNoteRequest": {
        "type": "object",
        "required": [
          "kind",
          "title"
        ],
        "properties": {
          "body": {
            "type": "string"
          },
          "embedding": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "Pre-computed embedding vector from the client. Optional: if omitted and the\nserver's embedder is ready, the server embeds the entry.\nIf neither is available, the entry is stored without a vector (text search only)."
          },
          "kind": {
            "type": "string",
            "description": "Kind of memory entry: `decision`, `requirement`, `note`, `question`, `handoff`, `intent`."
          },
          "linked_files": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Source file paths this entry is linked to."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional tags for filtering."
          },
          "title": {
            "type": "string"
          }
        }
      },
      "AddNoteResponse": {
        "type": "object",
        "required": [
          "stored",
          "id"
        ],
        "properties": {
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConflictEntry"
            },
            "description": "Conflicting entries (only present on 409)."
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the created note."
          },
          "stored": {
            "type": "boolean",
            "description": "Whether the note was stored (always true for 201/409)."
          }
        }
      },
      "BatchItemResult": {
        "type": "object",
        "description": "Per-entry outcome in the batch response.",
        "required": [
          "status",
          "external_id"
        ],
        "properties": {
          "external_id": {
            "type": "string"
          },
          "id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The note's `sync_id` (the same id `GET /memory/since` returns and\ncursors on, never the raw row id). Present for `\"created\"` and also\nfor a `\"skipped\"` dedupe-hit (the already-existing id), so a caller\nthat lost track of an earlier create can still recover the id from a\nplain re-push, and a caller that stamps this as its pull cursor gets\nan id that actually orders against `/memory/since`."
          },
          "status": {
            "type": "string",
            "description": "`\"created\"` or `\"skipped\"` (idempotent re-push)."
          }
        }
      },
      "BatchNoteItem": {
        "type": "object",
        "description": "One entry in a `POST /memory/batch` request. Field-for-field match of the\nCLI's `BatchPushItem` (`inkentry-core/src/storage/remote/sync.rs`): the CLI\nnever sends `embedding` today (its push is text-only), but the\nfield is accepted when present for forward compatibility with a possible\nfuture pushed-vector optimization: the server must not require it.",
        "required": [
          "kind",
          "title",
          "external_id"
        ],
        "properties": {
          "body": {
            "type": [
              "string",
              "null"
            ]
          },
          "embedding": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "Optional pre-computed embedding. Tolerated, never required."
          },
          "external_id": {
            "type": "string",
            "description": "Stable cross-machine identity: the server's idempotency key. Stored as\n`notes.remote_id` (unique). A re-push of the same `external_id` against\na live note is skipped, not duplicated."
          },
          "kind": {
            "type": "string"
          },
          "source_commit": {
            "type": [
              "string",
              "null"
            ],
            "description": "Git SHA provenance, if any. No dedicated column on this schema; stored\nas a `git:<sha>` tag, the same convention `harvested_shas` already reads."
          },
          "title": {
            "type": "string"
          }
        }
      },
      "BatchPushRequest": {
        "type": "object",
        "required": [
          "entries"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchNoteItem"
            }
          }
        }
      },
      "BatchPushResponse": {
        "type": "object",
        "description": "Response body for `POST /memory/batch`; always `207 Multi-Status`.",
        "required": [
          "created",
          "skipped",
          "failed",
          "results"
        ],
        "properties": {
          "created": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "failed": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchItemResult"
            }
          },
          "skipped": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ConflictEntry": {
        "type": "object",
        "description": "A single conflicting memory entry returned in a 409 response.",
        "required": [
          "id",
          "title",
          "similarity"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "similarity": {
            "type": "number",
            "format": "float",
            "description": "Cosine similarity to the new entry (0.0–1.0)."
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Consistent JSON error body: `{\"error\": {\"code\": \"...\", \"message\": \"...\"}}`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "HarvestedShasResponse": {
        "type": "object",
        "description": "Object envelope for the harvested-SHAs read endpoint.\n\nA JSON response root must be an object, never a bare array of primitives\n(ADR-076: the memory wire contract).",
        "required": [
          "shas"
        ],
        "properties": {
          "shas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "NoteListResponse": {
        "type": "object",
        "description": "Object envelope for the list and search read endpoints.\n\nA JSON response root must be an object, never a bare array (ADR-076: the\nmemory wire contract). `entries` carries the notes; `total` is their count\nin this response, which is already `limit`-capped, not a project-wide count\n(that has its own `/stats` route).",
        "required": [
          "entries",
          "total"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerNote"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "ServerNote": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "body",
          "tags",
          "linked_files",
          "created_at",
          "status"
        ],
        "properties": {
          "body": {
            "type": "string"
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of creation."
          },
          "distance": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Cosine distance from query (only present in search results)."
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kind": {
            "type": "string",
            "description": "Kind: `decision`, `requirement`, `note`, `question`, `handoff`, or `intent`."
          },
          "linked_files": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "remote_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Canonical cross-machine id (uuid). Optional and additive; `None` for\nrows never assigned one. Absent on the wire when `None`."
          },
          "status": {
            "type": "string",
            "description": "`active` or `archived`."
          },
          "superseded_by": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Pass as `Authorization: Bearer <key>`. Not required when no key is configured on the server."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}